# server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Custom Chatbot API
  description: |
    A comprehensive chatbot API that integrates with BotDojo for supplement discovery and health recommendations.
    
    ## Features
    - Real-time chat with AI assistant
    - Product recommendations with structured content
    - Suggested follow-up questions
    - Debug endpoints for development
    - Rate limiting and security features
    - Caching for improved performance
    
    ## Authentication
    BotDojo credentials must be provided in the request body as initData field (encrypted or plain JSON) for each API request.
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.example.com
    description: Production server
paths:
...
  /chat:
    post:
      summary: Send Chat Message
      description: Send a message to the chatbot and receive AI-generated responses with product recommendations
      tags:
        - Chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
                - initData
              properties:
                message:
                  type: string
                  minLength: 1
                  maxLength: 1000
                  description: The user's message to the chatbot
                  example: "What supplements can help with sleep?"
                initData:
                  type: string
                  description: BotDojo credentials (encrypted or plain JSON string). Must include BOTDOJO_API_KEY, BOTDOJO_BASE_URL, BOTDOJO_ACCOUNT_ID, BOTDOJO_PROJECT_ID, and BOTDOJO_FLOW_ID. Can be encrypted using RSA encryption or sent as plain JSON.
                  example: '{"BOTDOJO_API_KEY":"your-api-key","BOTDOJO_BASE_URL":"https://api.botdojo.com","BOTDOJO_ACCOUNT_ID":"account-id","BOTDOJO_PROJECT_ID":"project-id","BOTDOJO_FLOW_ID":"flow-id"}'
      responses:
        '200':
          description: Successful response with chatbot messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
                  debug:
                    $ref: '#/components/schemas/DebugInfo'
        '400':
          description: Bad request - invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many requests - rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /suggestions:
    post:
      summary: Get Suggested Questions
      description: Get suggested follow-up questions based on context
      tags:
        - Suggestions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - initData
              properties:
                context:
                  type: string
                  description: Context for generating suggestions
                  example: "sleep and relaxation"
                currentSetIndex:
                  type: integer
                  minimum: 0
                  description: Index of the current suggestion set
                  example: 0
                initData:
                  type: string
                  description: BotDojo credentials (encrypted or plain JSON string). Must include BOTDOJO_API_KEY, BOTDOJO_BASE_URL, BOTDOJO_ACCOUNT_ID, BOTDOJO_PROJECT_ID, and BOTDOJO_FLOW_ID. Can be encrypted using RSA encryption or sent as plain JSON.
                  example: '{"BOTDOJO_API_KEY":"your-api-key","BOTDOJO_BASE_URL":"https://api.botdojo.com","BOTDOJO_ACCOUNT_ID":"account-id","BOTDOJO_PROJECT_ID":"project-id","BOTDOJO_FLOW_ID":"flow-id"}'
      responses:
        '200':
          description: Successful response with suggested questions
          content:
            application/json:
              schema:
                type: object
                properties:
                  suggestedQuestions:
                    type: array
                    items:
                      type: array
                      items:
                        type: string
                    example: [["What supplements can help with sleep?", "What can I take for stress?", "How do I support my immune system?"]]
                  totalSets:
                    type: integer
                    example: 3
                  currentSetIndex:
                    type: integer
                    example: 0
        '400':
          description: Bad request - invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /debug-botdojo:
    post:
      summary: Debug BotDojo Response
      description: Get raw BotDojo API response for debugging purposes
      tags:
        - Debug
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
                - initData
              properties:
                message:
                  type: string
                  minLength: 1
                  maxLength: 1000
                  description: The message to send to BotDojo
                  example: "What supplements can help with sleep?"
                initData:
                  type: string
                  description: BotDojo credentials (encrypted or plain JSON string). Must include BOTDOJO_API_KEY, BOTDOJO_BASE_URL, BOTDOJO_ACCOUNT_ID, BOTDOJO_PROJECT_ID, and BOTDOJO_FLOW_ID. Can be encrypted using RSA encryption or sent as plain JSON.
                  example: '{"BOTDOJO_API_KEY":"your-api-key","BOTDOJO_BASE_URL":"https://api.botdojo.com","BOTDOJO_ACCOUNT_ID":"account-id","BOTDOJO_PROJECT_ID":"project-id","BOTDOJO_FLOW_ID":"flow-id"}'
      responses:
        '200':
          description: Raw BotDojo response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  rawResponse:
                    type: object
                    description: Raw response from BotDojo API
                  endpoint:
                    type: string
                    example: "https://api.botdojo.com/accounts/123/projects/456/flows/789/run"
                  requestBody:
                    type: object
                    description: Request body sent to BotDojo
  /test-structured:
    post:
      summary: Test Structured Content
      description: Generate test messages with different types of structured content
      tags:
        - Testing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - contentType
                - initData
              properties:
                contentType:
                  type: string
                  enum: [guide, faq, labResult, image, linkList, product]
                  description: Type of structured content to generate
                  example: "product"
                initData:
                  type: string
                  description: BotDojo credentials (encrypted or plain JSON string). Must include BOTDOJO_API_KEY, BOTDOJO_BASE_URL, BOTDOJO_ACCOUNT_ID, BOTDOJO_PROJECT_ID, and BOTDOJO_FLOW_ID. Can be encrypted using RSA encryption or sent as plain JSON.
                  example: '{"BOTDOJO_API_KEY":"your-api-key","BOTDOJO_BASE_URL":"https://api.botdojo.com","BOTDOJO_ACCOUNT_ID":"account-id","BOTDOJO_PROJECT_ID":"project-id","BOTDOJO_FLOW_ID":"flow-id"}'
      responses:
        '200':
          description: Test message with structured content
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
...
components:
  schemas:
    Message:
      type: object
      required:
        - id
        - role
        - type
        - content
      properties:
        id:
          type: string
          description: Unique identifier for the message
          example: "msg-1642248600000-abc123"
        role:
          type: string
          enum: [user, bot]
          description: Role of the message sender
          example: "bot"
        type:
          type: string
          enum: [text, buttons, card, list, typing, canvas, product]
          description: Type of message content
          example: "text"
        content:
          type: object
          description: Message content (structure varies by type)
          properties:
            text:
              type: string
              description: Text content for text messages
              example: "Here are some supplements that can help with sleep:"
            options:
              type: array
              items:
                type: string
              description: Button options for button messages
              example: ["Energy", "Immunity", "Vitamins"]
        suggestedQuestions:
          type: array
          items:
            type: string
          description: Suggested follow-up questions
          example: ["What supplements can help with sleep?", "What can I take for stress?"]
        structured:
          $ref: '#/components/schemas/StructuredContent'
    StructuredContent:
      type: object
      required:
        - type
        - data
      properties:
        type:
          type: string
          enum: [product, guide, faq, labResult, image, linkList]
          description: Type of structured content
          example: "product"
        data:
          type: array
          items:
            type: object
          description: Array of structured content items
          example: [
            {
              "sku": "MAG-001",
              "productId": "12345",
              "title": "Magnesium Glycinate",
              "imageUrl": "https://example.com/image.jpg",
              "description": "High-quality magnesium for better sleep",
              "url": "https://example.com/product/mag-001"
            }
          ]
    DebugInfo:
      type: object
      properties:
        rawBotDojoResponse:
          type: object
          description: Raw response from BotDojo API
        endpoint:
          type: string
          description: BotDojo API endpoint used
          example: "https://api.botdojo.com/accounts/123/projects/456/flows/789/run"
        requestBody:
          type: object
          description: Request body sent to BotDojo
        cached:
          type: boolean
          description: Whether the response was served from cache
          example: false
    Error:
      type: object
      required:
        - error
        - statusCode
      properties:
        error:
          type: string
          description: Error message
          example: "Message is required and must be a non-empty string"
        statusCode:
          type: integer
          description: HTTP status code
          example: 400
        requestId:
          type: string
          description: Unique request identifier for debugging
          example: "req_1642248600000_abc123"
        stack:
          type: string
          description: Error stack trace (only in development)
          example: "Error: Message is required\n    at validateString (/app/src/utils/errorHandler.ts:45:11)"

  # Note: BotDojo credentials are provided in request body as initData field
  # (encrypted or plain JSON) rather than as a header
...